fix : 트랜잭션 매니저, DB분리 수정
