+++ src/main/java/com/deliveryfood/config/DataSourceConfig.java
@@ -1,22 +1,16 @@
 package com.deliveryfood.config;
 
+import com.deliveryfood.datasource.ReplicationRoutingDataSource;
 import com.zaxxer.hikari.HikariDataSource;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.ibatis.session.SqlSessionFactory;
-import org.mybatis.spring.SqlSessionFactoryBean;
-import org.mybatis.spring.SqlSessionTemplate;
-import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
 import org.springframework.jdbc.datasource.LazyConnectionDataSourceProxy;
-import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 import javax.sql.DataSource;
@@ -27,7 +21,6 @@ import java.util.Map;
 @Configuration
 @RequiredArgsConstructor
 @EnableTransactionManagement
-@MapperScan(basePackages = "com.deliveryfood.mapper")
 public class DataSourceConfig {
 
     @Primary
@@ -61,46 +54,11 @@ public class DataSourceConfig {
         routingDataSource.setDefaultTargetDataSource(masterDataSource);
 
         return routingDataSource;
-
-//        Map<Object, Object> targetDataSourcesMap = new HashMap<>();
-//        targetDataSourcesMap.put("master",masterDataSource);
-//        targetDataSourcesMap.put("slave", slaveDataSource);
-//
-//        return new ReplicationRoutingDataSource(targetDataSourcesMap);
     }
 
     @Bean(name = "proxyDataSource")
     public DataSource proxyDataSource(@Qualifier("routingDataSource") DataSource routingDataSource) {
         return new LazyConnectionDataSourceProxy(routingDataSource);
     }
-
-    @Bean
-    public PlatformTransactionManager transactionManager(@Qualifier(value = "proxyDataSource") DataSource dataSource) {
-        return new DataSourceTransactionManager(dataSource);
-    }
-
-
-    private final ApplicationContext applicationContext;
-
-    @Bean
-    public SqlSessionFactory sqlSession(DataSource dataSource) throws Exception {
-        log.info("###### TESTLOG sqlSession");
-        SqlSessionFactoryBean sqlSessionFactoryBean = new SqlSessionFactoryBean();
-        sqlSessionFactoryBean.setDataSource(dataSource);
-        sqlSessionFactoryBean.setMapperLocations(applicationContext.getResources("classpath:/mybatis/**/*.xml"));
-        return sqlSessionFactoryBean.getObject();
-    }
-
-    @Bean
-    public SqlSessionTemplate sqlSessionTemplate(SqlSessionFactory sqlSessionFactory) throws Exception {
-        log.info("###### TESTLOG sqlSessionTemplate");
-        return new SqlSessionTemplate(sqlSessionFactory);
-    }
-//
-//    @Bean(name = "transactionManager")
-//    public PlatformTransactionManager transactionManager(@Qualifier(value = "proxyDataSource") DataSource dataSource) {
-//        log.info("###### TESTLOG transactionManager");
-//        return new DataSourceTransactionManager(dataSource);
-//    }
 
 }
