+++ src/main/java/com/deliveryfood/dao/MenuDao.java
@@ -4,6 +4,7 @@ import com.deliveryfood.dto.MenuDto;
 import lombok.RequiredArgsConstructor;
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
