+++ src/main/java/com/deliveryfood/service/MenuService.java
@@ -16,7 +16,6 @@ public class MenuService implements IMenuService {
     private final MenuMapper menuMapper;
 
     @Override
-    @Transactional
     public void createMenuById(MenuInput menuInput) {
         MenuDto menuDto = MenuDto.builder()
                 .menuId(menuInput.getMenuId())
@@ -24,12 +23,9 @@ public class MenuService implements IMenuService {
                 .name(menuInput.getName())
                 .build();
         menuMapper.createMenuById(menuDto);
-
-        throw new RuntimeException();
     }
 
     @Override
-    @Transactional(readOnly = true)
     public MenuDto findMenus(MenuInput menuInput) {
         MenuDto menuDto = MenuDto.builder()
                 .menuId(menuInput.getMenuId())
@@ -39,7 +35,6 @@ public class MenuService implements IMenuService {
     }
 
     @Override
-    @Transactional(readOnly = true)
     public List<MenuDto> findMenuById(MenuInput menuInput) {
         MenuDto menuDto = MenuDto.builder()
                 .restaurantId(menuInput.getRestaurantId())
