+++ src/main/resources/application-dev.properties
@@ -1,30 +1,13 @@
 #db
-#spring.datasource.master.drive-class-name=com.mysql.cj.jdbc.Driver
-#spring.datasource.master.jdbc-url=jdbc:h2:tcp://localhost/~/test
-#spring.datasource.master.username=sa
-#spring.datasource.master.driver-class-name=org.h2.Driver
-
-spring.datasource.master.jdbc-url=jdbc:mysql://localhost:3306/TEST1
+spring.datasource.master.drive-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.master.jdbc-url=jdbc:h2:tcp://localhost/~/test
 spring.datasource.master.username=sa
-spring.datasource.master.password=password
-spring.datasource.master.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.master.driver-class-name=org.h2.Driver
 
-
-#spring.datasource.slave.drive-class-name=com.mysql.cj.jdbc.Driver
-#spring.datasource.slave.jdbc-url=jdbc:h2:tcp://localhost/~/test2
-#spring.datasource.slave.username=sa
-#spring.datasource.slave.driver-class-name=org.h2.Driver
-#spring.datasource.slave.drive-class-name=com.mysql.cj.jdbc.Driver
-
-#spring.datasource.slave.jdbc-url=jdbc:mysql://localhost:3306/TEST1
-#spring.datasource.slave.username=sa
-#spring.datasource.slave.password=password
-#spring.datasource.slave.driver-class-name=com.mysql.cj.jdbc.Driver
-
-spring.datasource.slave.drive-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.slave.jdbc-url=jdbc:h2:tcp://localhost/~/test
+spring.datasource.slave.jdbc-url=jdbc:mysql://localhost:3306/TEST1
 spring.datasource.slave.username=sa
-spring.datasource.slave.driver-class-name=org.h2.Driver
+spring.datasource.slave.password=password
+spring.datasource.slave.driver-class-name=com.mysql.cj.jdbc.Driver
 
 
 # mybatis
